pythonapp/requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
PyMySQL==1.0.3
SQLAlchemy==2.0.16
typing-extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0
